Reworked some of the validation code
